requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
sentence-transformers>=2.7.0
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.3.0
python-dotenv>=1.0.0
email-validator>=2.1.0
requests>=2.32.0
beautifulsoup4>=4.12.0
selenium>=4.20.0
matplotlib>=3.8.0
